requirements.txt
# Add Python dependencies for ClinicalBench here.

datasets==2.19.0
openai==1.30.0
anthropic==0.25.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.0
tqdm==4.66.4
python-dotenv==1.0.1
jsonlines==4.0.0
matplotlib==3.9.0
seaborn==0.13.2